Test_Project_Iqbal/pyproject.toml
[project]
name = "test-project-iqbal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
]
